fix: page components moved into pages directory
