+++ src/app/components/app/app.component.spec.ts
@@ -3,12 +3,12 @@ import { AngularFirestore } from '@angular/fire/firestore';
 import { Router } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
 import { APP_CONFIG, APP_UNIT_TEST_CONFIG } from '../../app-config';
+import { NotFoundComponent } from '../../pages/not-found/not-found.component';
 import { AlertService, PaginationService, SeoService } from '../../services';
 import { AlertComponent } from '../alert/alert.component';
 import { FooterComponent } from '../footer/footer.component';
 import { LoadingSpinnerComponent } from '../loading-spinner/loading-spinner.component';
 import { NavMenuComponent } from '../nav-menu/nav-menu.component';
-import { NotFoundComponent } from '../not-found/not-found.component';
 import { AppComponent } from './app.component';
 
 describe('AppComponent', () => {
