+++ src/app/pages/blog-detail/blog-detail.component.scss
