+++ src/app/pages/blog-detail/blog-detail.component.ts
