+++ src/app/pages/blog-list/blog-list.component.scss
