+++ src/app/pages/joke-detail/joke-detail.component.ts
