+++ src/app/pages/joke-list/joke-list.component.ts
