+++ src/app/pages/quote-detail/quote-detail.component.html
