+++ src/app/pages/quote-list/quote-list.component.ts
