+++ src/app/pages/search/search.component.scss
