+++ src/app/pages/search/search.component.spec.ts
@@ -2,8 +2,8 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormsModule } from '@angular/forms';
 import { RouterTestingModule } from '@angular/router/testing';
 import { APP_CONFIG, APP_UNIT_TEST_CONFIG } from '../../app-config';
+import { SideBarComponent } from '../../components/side-bar/side-bar.component';
 import { AlertService, SeoService } from '../../services';
-import { SideBarComponent } from '../side-bar/side-bar.component';
 import { SearchComponent } from './search.component';
 
 describe('SearchComponent', () => {
