+++ src/app/pages/taxonomy/taxonomy.component.scss
